Animations are working!
Or at least working the way I want them too for the most part. There is an idle animation, a jump animation that plays once (no repeat), and walking animations.. all working. The trick is we have to be very careful with the logic that plays animations. We can redo sprite atlas images to create more realisitic movements.